refactor: structure
