+++ src/routes/graphql/query-types/profile-queries.ts
@@ -1,7 +1,7 @@
 import { GraphQLNonNull } from 'graphql';
 import { Context } from '../basic-types/context.js';
 import { UUIDType } from '../basic-types/uuid.js';
-import { NonNullListOfNonNull } from '../basic-types/nonNull.js';
+import { NonNullListOfNonNull } from '../basic-types/non-null.js';
 import { Profile } from '../basic-types/profile.js';
 
 export const profilesQuery = {
